RAG/web scrapping/requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
trafilatura
playwright
langchain
openai
sentence-transformers
faiss-cpu
numpy
pydantic
python-dotenv
requests
tqdm
gunicorn
